refactor(ASKFLOOR-1-T3a): domain additions move to domain/human-decision.ts; types.ts stays at its 740-line ceiling
Ticket: ASKFLOOR-1/ASKFLOOR-1-T3a
